requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27
python-magic-bin==0.4.14
openai==1.3.7
spacy==3.7.2
language-tool-python==2.7.1
reportlab==4.0.7